// packages/sdk/src/contracts/generated.ts
export interface ContractFunctionDefinition {
  name: string;
  id: number;
  inputStruct?: string;
  outputStruct?: string;
}

export interface ContractDefinition {
  name: string;
  contractIndex: number;
  address: string;
  functions: readonly ContractFunctionDefinition[];
  procedures: readonly ContractFunctionDefinition[];
}
// ...
export type ContractName = typeof CONTRACT_DEFINITIONS[number]['name'];

type DefinitionByName<TName extends ContractName> = Extract<typeof CONTRACT_DEFINITIONS[number], { name: TName }>;

export type ContractFunctionName<TName extends ContractName> = DefinitionByName<TName>['functions'][number]['name'];
export type ContractProcedureName<TName extends ContractName> = DefinitionByName<TName>['procedures'][number]['name'];

export const CONTRACT_DEFINITION_MAP = new Map<ContractName, ContractDefinition>(CONTRACT_DEFINITIONS.map((definition) => [definition.name, definition]));
export const CONTRACT_DEFINITION_BY_INDEX = new Map<number, ContractDefinition>(CONTRACT_DEFINITIONS.map((definition) => [definition.contractIndex, definition]));
export const CONTRACT_DEFINITION_BY_ADDRESS = new Map<string, ContractDefinition>(CONTRACT_DEFINITIONS.map((definition) => [definition.address, definition]));